break out into PanelPreview component
